Separate UI and Controller classes
