+++ TrainController/TrainController.sln
@@ -3,7 +3,12 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.2.32616.157
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TrainController", "TrainController\TrainController.csproj", "{ED89FEB5-D9FA-4077-B111-E5AFD4AA6135}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TrainController", "TrainController\TrainController.csproj", "{ED89FEB5-D9FA-4077-B111-E5AFD4AA6135}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{ED838E99-4B89-4E4E-A861-D4E2691A9A66}"
+	ProjectSection(SolutionItems) = preProject
+		.editorconfig = .editorconfig
+	EndProjectSection
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
