+++ TrainController/TrainController/TrainDriverInterface.xaml
@@ -13,21 +13,15 @@
 
         <TextBlock x:Name="CmdAuthority" HorizontalAlignment="Left" Margin="41,160,0,0" TextWrapping="Wrap" Text="Commanded Authority:&#xa;0 blocks" VerticalAlignment="Top" Height="60" Width="218" Background="#FF64E164" FontSize="20" TextAlignment="Center"/>
         <TextBlock x:Name="CurAuthority" HorizontalAlignment="Left" Margin="41,237,0,0" TextWrapping="Wrap" Text="Current Authority:&#xa;0 blocks" VerticalAlignment="Top" Height="60" Width="218" Background="#FF64E164" FontSize="20" TextAlignment="Center"/>
-        <!--<TextBlock HorizontalAlignment="Left" Margin="41,237,0,0" TextWrapping="Wrap" VerticalAlignment="Top" Height="60" Width="218" Background="#FF64E164" FontSize="20" TextAlignment="Center"><Run Text="Current Authority:"/><LineBreak/><Run/></TextBlock>-->
-        <!--<TextBlock HorizontalAlignment="Left" Margin="41,75,0,0" TextWrapping="Wrap" Text="Service Brakes:" VerticalAlignment="Top" Height="70" Width="218" Background="#FFFF5A5A" FontSize="20" TextAlignment="Center"/>-->
+
         <TextBlock x:Name="CurSpeed" HorizontalAlignment="Left" Margin="280,138,0,0" TextWrapping="Wrap" Text="Current Speed:&#xa;0 mph" VerticalAlignment="Top" Height="119" Width="242" Background="#FF64E164" FontSize="34" TextAlignment="Center"/>
         <TextBlock x:Name="CmdSpeed" HorizontalAlignment="Left" Margin="280,271,0,0" TextWrapping="Wrap" Text="Cmd Speed:&#xa;0 mph" VerticalAlignment="Top" Height="62" Width="109" Background="#FF64E164" FontSize="18" TextAlignment="Center"/>
         <TextBlock x:Name="SetSpeed" HorizontalAlignment="Left" Margin="410,271,0,0" TextWrapping="Wrap" Text="Set Speed:&#xa;        mph" VerticalAlignment="Top" Height="62" Width="109" Background="#FFFFDF20" FontSize="18" TextAlignment="Center"/>
         <TextBlock x:Name="CurPower" HorizontalAlignment="Left" Margin="282,345,0,0" TextWrapping="Wrap" Text="Power: 0 kW" VerticalAlignment="Top" Height="32" Width="238" Background="#FF64E164" FontSize="18" TextAlignment="Center"/>
         <TextBlock x:Name="Beacon" HorizontalAlignment="Left" Margin="540,317,0,0" TextWrapping="Wrap" Text="Nearest Beacon:&#xa;-" VerticalAlignment="Top" Height="60" Width="220" Background="#FF64E164" FontSize="20" TextAlignment="Center"/>
         <TextBlock x:Name="Temperature" HorizontalAlignment="Left" Margin="540,249,0,0" TextWrapping="Wrap" Text="Temperature: 72°F" VerticalAlignment="Top" Height="54" Width="109" Background="#FF57F1D9" FontSize="16" TextAlignment="Center"/>
-        <!--<TextBlock HorizontalAlignment="Left" Margin="102,268,0,0" TextWrapping="Wrap" Text="0 blocks" VerticalAlignment="Top" Background="White" Height="18" Width="96" TextAlignment="Center"/>
-        <TextBlock HorizontalAlignment="Left" Margin="102,195,0,0" TextWrapping="Wrap" Text="0 blocks" VerticalAlignment="Top" Background="White" Height="16" Width="96" TextAlignment="Center"/>-->
-        <!--<TextBlock HorizontalAlignment="Left" Margin="594,348,0,0" TextWrapping="Wrap" Text="Mt. Washington" VerticalAlignment="Top" Background="White" TextAlignment="Center" Width="111"/>-->
-        <!--<TextBlock x:Name="Power" HorizontalAlignment="Left" Margin="410,352,0,0" TextWrapping="Wrap" Text="0 kW" VerticalAlignment="Top" Background="White" Width="90" TextAlignment="Center"/>-->
-        <TextBlock x:Name="SelectType" HorizontalAlignment="Left" Margin="41,392,0,0" TextWrapping="Wrap" Text="Select Controller Type" VerticalAlignment="Top" Height="27" Width="218" Background="#FFDADADA" FontSize="16" TextAlignment="Center"/>
 
-        <!--<Slider x:Name="ServiceBrake" HorizontalAlignment="Left" Margin="50,110,0,0" VerticalAlignment="Top" Width="200" Height="20"/>-->
+        <TextBlock x:Name="SelectType" HorizontalAlignment="Left" Margin="41,392,0,0" TextWrapping="Wrap" Text="Select Controller Type" VerticalAlignment="Top" Height="27" Width="218" Background="#FFDADADA" FontSize="16" TextAlignment="Center"/>
 
         <Button x:Name="AutoMode" Click="Button_Click" ClickMode="Press" Content="Automatic Mode" HorizontalAlignment="Left" VerticalAlignment="Top" Margin="279,77,0,0" Height="43" Width="110"/>
         <Button x:Name="ManualMode" Click="Button_Click" ClickMode="Press" Content="Manual Mode" HorizontalAlignment="Left" VerticalAlignment="Top" Margin="408,77,0,0" Height="43" Width="110"/>
@@ -43,9 +37,7 @@
 
         <Button x:Name="EngineerPanel" Click="Button_Click" ClickMode="Press" Content="Open Engineer's Panel" HorizontalAlignment="Left" VerticalAlignment="Top" Margin="282,392,0,0" Height="28" Width="240" FontSize="16"/>
         <Button x:Name="TestPanel" Click="Button_Click" ClickMode="Press" Content="Open Test Panel" HorizontalAlignment="Left" VerticalAlignment="Top" Margin="540,392,0,0" Height="28" Width="220" FontSize="16"/>
-
-        <!--<TextBlock HorizontalAlignment="Left" Margin="290,303,0,0" TextWrapping="Wrap" Text="0 mph" VerticalAlignment="Top" Background="White" Height="18" Width="88" TextAlignment="Center"/>-->
-        <!--<TextBlock HorizontalAlignment="Left" Margin="335,195,0,0" TextWrapping="Wrap" Text="0 mph" VerticalAlignment="Top" Background="White" Height="50" Width="125" FontSize="28" TextAlignment="Center"/>-->
+        
         <TextBox x:Name="SetSpeedBox" HorizontalAlignment="Left" Margin="422,302,0,0" TextWrapping="Wrap" Text="0" VerticalAlignment="Top" Width="41" TextAlignment="Center" KeyDown="KeyDownButton"/>
 
     </Grid>
