+++ TrainController/TrainController/TrainDriverInterface.xaml.cs
@@ -22,45 +22,22 @@ namespace TrainController
     /// </summary>
     public partial class MainWindow : Window
     {
-        // Serial port for connecting to Raspberry Pi:
-        SerialPort pi = new SerialPort();
+        public List<Controller> mTrainSet;
+        public Controller mSelectedTrain;
 
-        // Boolean for switching between auto and manual driving modes:
-        // 'false' is software controller, 'true' is hardware controller:
-        public bool mControlType;
-
-        public bool mAutoMode = false;
-        public bool mLeftDoorsStatus = false;
-        public bool mRightDoorsStatus = false;
-        public bool mInteriorLightsStatus = false;
-        public bool mExteriorLightsStatus = false;
-        public bool mAnnouncementsStatus = false;
-        public bool mServiceBrakeStatus = false;
-        public bool mEmergencyBrakeStatus = false;
-
-        public int mTemperature = 72;
-        public int mKp = 0;
-        public int mKi = 0;
-        public int mCurSpeed = 0;
-        public int mCmdSpeed = 0;
-        public int mSetSpeed = 0;
-        public int mCmdAuthority = 0;
-        public int mCurAuthority = 0;
-        public string mBeacon = "-";
-        public float mCurPower = 0;
-
-        public const float Pmax = 120000; // 120 kW
-        public float Uk = 0;
         public int T = 250; // 250 ms
 
         public MainWindow()
         {
             InitializeComponent();
 
+            mSelectedTrain = new Controller();
+            //mTrainSet[0] = mSelectedTrain;
+
             // Disable all buttons on main window until a HW_SW window option is selected:
             ManualMode.IsEnabled = false;
             AutoMode.IsEnabled = false;
-            ServiceBrake.IsEnabled = false;
+            ServiceBrake.IsEnabled = true;
             EmergencyBrake.IsEnabled = false;
             SetSpeedBox.IsEnabled = false;
             EngineerPanel.IsEnabled = false;
@@ -73,28 +50,10 @@ namespace TrainController
             InteriorLights.IsEnabled = false;
             ExteriorLights.IsEnabled = false;
 
-            InitTimer();
-
             HW_SW selectType = new HW_SW();
             selectType.Topmost = true;
             selectType.Show();
             selectType.Activate();
-        }
-
-        public void setupHardware()
-        {
-            // Setup serial port information: 
-            pi.PortName = "COM3";
-            pi.BaudRate = 115200;
-            pi.Parity = Parity.None;
-            pi.DataBits = 8;
-            pi.StopBits = StopBits.One;
-            pi.Handshake = Handshake.None;
-            pi.WriteTimeout = 500;
-
-            // Open serial port and reset all stored data:
-            pi.Open();
-            //pi.WriteLine("?");
         }
 
         public void Button_Click(object sender, RoutedEventArgs e)
@@ -115,18 +74,11 @@ namespace TrainController
                 SetSpeedBox.IsEnabled = false;
 
                 SetSpeed.Background = new SolidColorBrush(Color.FromArgb(0x30, 0, 0, 0));
-                SetSpeedBox.Text = mSetSpeed.ToString();
+                SetSpeedBox.Text = mSelectedTrain.mSetSpeed.ToString();
 
-                mSetSpeed = mCmdSpeed;
-                mAutoMode = true;
-
-                // Hardware Controls:
-                if (mControlType)
-                {
-                    pi.WriteLine("m");
-                    string output = pi.ReadLine();
-                }
+                mSelectedTrain.setAutoMode();
             }
+
             else if (sender == ManualMode)
             {
                 // UI Element Controls:
@@ -144,275 +96,116 @@ namespace TrainController
 
                 SetSpeed.Background = new SolidColorBrush(Color.FromArgb(0xFF, 0xFF, 0xDF, 0x20));
 
-                mAutoMode = false;
-
-                // Hardware Controls:
-                if (mControlType)
-                {
-                    pi.WriteLine("m");
-                    string output = pi.ReadLine();
-                }
+                mSelectedTrain.setManualMode();
             }
 
             else if (sender == EmergencyBrake)
             {
-                if (!mEmergencyBrakeStatus)
+                if (!mSelectedTrain.mEmergencyBrakeStatus)
                 {
-                    mEmergencyBrakeStatus = true;
                     EmergencyBrake.Content = "Emergency Brake\n         (ON)";
                     EmergencyBrake.Background = new SolidColorBrush((Color)ColorConverter.ConvertFromString("#FF0000"));
-                }
 
-                // Hardware Controls:
-                if (mControlType)
-                {
-                    pi.WriteLine("e");
-                    string output = pi.ReadLine();
+                    mSelectedTrain.setEmergencyBrake();
                 }
             }
 
             else if (sender == ServiceBrake)
             {
-                if(!mControlType)
-                {
-                    if (!mServiceBrakeStatus)
-                    {
-                        mServiceBrakeStatus = true;
-                        ServiceBrake.Content = "Service Brake\n      (ON)";
-                        ServiceBrake.Background = new SolidColorBrush((Color)ColorConverter.ConvertFromString("#FF0000"));
-                    }
-                    else
-                    {
-                        mServiceBrakeStatus = false;
-                        ServiceBrake.Content = "Service Brake\n      (OFF)";
-                        ServiceBrake.Background = new SolidColorBrush((Color)ColorConverter.ConvertFromString("#FFFF5A5A"));
-                    }
-                }
+                mSelectedTrain.setServiceBrake();
 
-                // Hardware  Controls:
+                if (mSelectedTrain.mServiceBrakeStatus)
+                {
+                    ServiceBrake.Content = "Service Brake\n      (ON)";
+                    ServiceBrake.Background = new SolidColorBrush((Color)ColorConverter.ConvertFromString("#FF0000"));
+                }
                 else
                 {
-                    pi.WriteLine("u");
-                    string output = pi.ReadLine();
-
-                    if (output == "On")
-                    {
-                        mServiceBrakeStatus = true;
-                        ServiceBrake.Content = "Service Brake\n      (ON)";
-                        ServiceBrake.Background = new SolidColorBrush((Color)ColorConverter.ConvertFromString("#FF0000"));
-                    }
-                    else
-                    {
-                        mServiceBrakeStatus = false;
-                        ServiceBrake.Content = "Service Brake\n      (OFF)";
-                        ServiceBrake.Background = new SolidColorBrush((Color)ColorConverter.ConvertFromString("#FFFF5A5A"));
-                    }
+                    ServiceBrake.Content = "Service Brake\n      (OFF)";
+                    ServiceBrake.Background = new SolidColorBrush((Color)ColorConverter.ConvertFromString("#FFFF5A5A"));
                 }
             }
 
             else if (sender == LeftDoors)
             {
-                if (!mControlType)
+                mSelectedTrain.setLeftDoors();
+
+                if (mSelectedTrain.mLeftDoorsStatus)
                 {
-                    if (!mLeftDoorsStatus)
-                    {
-                        mLeftDoorsStatus = true;
-                        LeftDoors.Content = "Doors - Left\n    (OPEN)";
-                    }
-                    else
-                    {
-                        mLeftDoorsStatus = false;
-                        LeftDoors.Content = "Doors - Left\n   (CLOSED)";
-                    }
+                    LeftDoors.Content = "Doors - Left\n    (OPEN)";
                 }
                 else
                 {
-                    pi.WriteLine("l");
-                    string output = pi.ReadLine();
-
-                    if (output == "Open")
-                    {
-                        mLeftDoorsStatus = true;
-                        LeftDoors.Content = "Doors - Left\n    (OPEN)";
-                    }
-                    else
-                    {
-                        mLeftDoorsStatus = false;
-                        LeftDoors.Content = "Doors - Left\n   (CLOSED)";
-                    }
+                    LeftDoors.Content = "Doors - Left\n   (CLOSED)";
                 }
             }
 
             else if (sender == RightDoors)
             {
-                if (!mControlType)
+                mSelectedTrain.setRightDoors();
+
+                if (mSelectedTrain.mRightDoorsStatus)
                 {
-                    if (!mRightDoorsStatus)
-                    {
-                        mRightDoorsStatus = true;
-                        RightDoors.Content = "Doors - Right\n    (OPEN)";
-                    }
-                    else
-                    {
-                        mRightDoorsStatus = false;
-                        RightDoors.Content = "Doors - Right\n   (CLOSED)";
-                    }
+                    RightDoors.Content = "Doors - Right\n    (OPEN)";
                 }
                 else
                 {
-                    pi.WriteLine("r");
-                    string output = pi.ReadLine();
-
-                    if (output == "Open")
-                    {
-                        mRightDoorsStatus = true;
-                        RightDoors.Content = "Doors - Right\n    (OPEN)";
-                    }
-                    else
-                    {
-                        mRightDoorsStatus = false;
-                        RightDoors.Content = "Doors - Right\n   (CLOSED)";
-                    }
+                    RightDoors.Content = "Doors - Right\n   (CLOSED)";
                 }
             }
 
             else if (sender == InteriorLights)
             {
-                if (!mControlType)
+                mSelectedTrain.setInteriorLights();
+
+                if (mSelectedTrain.mInteriorLightsStatus)
                 {
-                    if (!mInteriorLightsStatus)
-                    {
-                        mInteriorLightsStatus = true;
-                        InteriorLights.Content = "Lights - Interior\n        (ON)";
-                    }
-                    else
-                    {
-                        mInteriorLightsStatus = false;
-                        InteriorLights.Content = "Lights - Interior\n        (OFF)";
-                    }
+                    InteriorLights.Content = "Lights - Interior\n        (ON)";
                 }
                 else
                 {
-                    pi.WriteLine("i");
-                    string output = pi.ReadLine();
-
-                    if (output == "On")
-                    {
-                        mInteriorLightsStatus = true;
-                        InteriorLights.Content = "Lights - Interior\n        (ON)";
-                    }
-                    else
-                    {
-                        mInteriorLightsStatus = false;
-                        InteriorLights.Content = "Lights - Interior\n        (OFF)";
-                    }
+                    InteriorLights.Content = "Lights - Interior\n        (OFF)";
                 }
             }
 
             else if (sender == ExteriorLights)
             {
-                if (!mControlType)
+                mSelectedTrain.setExteriorLights();
+
+                if (mSelectedTrain.mExteriorLightsStatus)
                 {
-                    if (!mExteriorLightsStatus)
-                    {
-                        mExteriorLightsStatus = true;
-                        ExteriorLights.Content = "Lights - Exterior\n        (ON)";
-                    }
-                    else
-                    {
-                        mExteriorLightsStatus = false;
-                        ExteriorLights.Content = "Lights - Exterior\n        (OFF)";
-                    }
+                    ExteriorLights.Content = "Lights - Exterior\n        (ON)";
                 }
                 else
                 {
-                    pi.WriteLine("x");
-                    string output = pi.ReadLine();
-
-                    if (output == "On")
-                    {
-                        mExteriorLightsStatus = true;
-                        ExteriorLights.Content = "Lights - Exterior\n        (ON)";
-                    }
-                    else
-                    {
-                        mExteriorLightsStatus = false;
-                        ExteriorLights.Content = "Lights - Exterior\n        (OFF)";
-                    }
+                    ExteriorLights.Content = "Lights - Exterior\n        (OFF)";
                 }
             }
 
             else if (sender == Announcements)
             {
-                if (!mControlType)
+                if (mSelectedTrain.mAnnouncementsStatus)
                 {
-                    if (!mAnnouncementsStatus)
-                    {
-                        mAnnouncementsStatus = true;
-                        Announcements.Content = "Announcements\n        (ON)";
-                    }
-                    else
-                    {
-                        mAnnouncementsStatus = false;
-                        Announcements.Content = "Announcements\n        (OFF)";
-                    }
+                    Announcements.Content = "Announcements\n        (ON)";
                 }
                 else
                 {
-                    pi.WriteLine("a");
-                    string output = pi.ReadLine();
-
-                    if (output == "On")
-                    {
-                        mAnnouncementsStatus = true;
-                        Announcements.Content = "Announcements\n        (ON)";
-                    }
-                    else
-                    {
-                        mAnnouncementsStatus = false;
-                        Announcements.Content = "Announcements\n        (OFF)";
-                    }
+                    Announcements.Content = "Announcements\n        (OFF)";
                 }
             }
 
             else if (sender == TempIncrease)
             {
-                if (!mControlType)
-                {
-                    mTemperature++;
-                }
-                else
-                {
-                    pi.WriteLine("h");
-                    string output = pi.ReadLine();
+                mSelectedTrain.tempIncrease();
 
-                    if(output == "tempIncreased")
-                    {
-                        mTemperature++;
-                    }
-                }
-
-                Temperature.Text = "Temperature: " + mTemperature.ToString() + "°F";
+                Temperature.Text = "Temperature: " + mSelectedTrain.mTemperature.ToString() + "°F";
             }
 
             else if (sender == TempDecrease)
             {
-                if (!mControlType)
-                {
-                    mTemperature--;
-                }
-                else
-                {
-                    pi.WriteLine("c");
-                    string output = pi.ReadLine();
+                mSelectedTrain.tempDecrease();
 
-                    if (output == "tempDecreased")
-                    {
-                        mTemperature--;
-                    }
-                }
-
-                Temperature.Text = "Temperature: " + mTemperature.ToString() + "°F";
+                Temperature.Text = "Temperature: " + mSelectedTrain.mTemperature.ToString() + "°F";
             }
 
             else if (sender == EngineerPanel)
@@ -420,8 +213,8 @@ namespace TrainController
                 EngineerPanel ePan = new EngineerPanel();
                 ePan.Owner = this;
 
-                ePan.Kp.Text = mKp.ToString();
-                ePan.Ki.Text = mKi.ToString();
+                ePan.Kp.Text = mSelectedTrain.mKp.ToString();
+                ePan.Ki.Text = mSelectedTrain.mKi.ToString();
 
                 ePan.Show();
             }
@@ -431,7 +224,7 @@ namespace TrainController
                 TestPanel tPan = new TestPanel();
                 tPan.Owner = this;
 
-                if (mLeftDoorsStatus)
+                if (mSelectedTrain.mLeftDoorsStatus)
                 {
                     tPan.LeftDoors.Content = "OPEN";
                     tPan.LeftDoors.Background = new SolidColorBrush((Color)ColorConverter.ConvertFromString("#FF70D060"));
@@ -442,7 +235,7 @@ namespace TrainController
                     tPan.LeftDoors.Background = new SolidColorBrush((Color)ColorConverter.ConvertFromString("#FFFF5050"));
                 }
 
-                if (mRightDoorsStatus)
+                if (mSelectedTrain.mRightDoorsStatus)
                 {
                     tPan.RightDoors.Content = "OPEN";
                     tPan.RightDoors.Background = new SolidColorBrush((Color)ColorConverter.ConvertFromString("#FF70D060"));
@@ -453,7 +246,7 @@ namespace TrainController
                     tPan.RightDoors.Background = new SolidColorBrush((Color)ColorConverter.ConvertFromString("#FFFF5050"));
                 }
 
-                if (mInteriorLightsStatus)
+                if (mSelectedTrain.mInteriorLightsStatus)
                 {
                     tPan.InteriorLights.Content = "ON";
                     tPan.InteriorLights.Background = new SolidColorBrush((Color)ColorConverter.ConvertFromString("#FF70D060"));
@@ -464,7 +257,7 @@ namespace TrainController
                     tPan.InteriorLights.Background = new SolidColorBrush((Color)ColorConverter.ConvertFromString("#FFFF5050"));
                 }
 
-                if (mExteriorLightsStatus)
+                if (mSelectedTrain.mExteriorLightsStatus)
                 {
                     tPan.ExteriorLights.Content = "ON";
                     tPan.ExteriorLights.Background = new SolidColorBrush((Color)ColorConverter.ConvertFromString("#FF70D060"));
@@ -475,7 +268,7 @@ namespace TrainController
                     tPan.ExteriorLights.Background = new SolidColorBrush((Color)ColorConverter.ConvertFromString("#FFFF5050"));
                 }
 
-                if (mAnnouncementsStatus)
+                if (mSelectedTrain.mAnnouncementsStatus)
                 {
                     tPan.Announcements.Content = "ON";
                     tPan.Announcements.Background = new SolidColorBrush((Color)ColorConverter.ConvertFromString("#FF70D060"));
@@ -486,16 +279,16 @@ namespace TrainController
                     tPan.Announcements.Background = new SolidColorBrush((Color)ColorConverter.ConvertFromString("#FFFF5050"));
                 }
 
-                tPan.Temperature.Text = mTemperature.ToString();
+                tPan.Temperature.Text = mSelectedTrain.mTemperature.ToString();
 
-                tPan.SetKp.Text = mKp.ToString();
-                tPan.SetKi.Text = mKi.ToString();
+                tPan.SetKp.Text = mSelectedTrain.mKp.ToString();
+                tPan.SetKi.Text = mSelectedTrain.mKi.ToString();
 
-                tPan.CurSpeed.Text = mCurSpeed.ToString();
-                tPan.CmdSpeed.Text = mCmdSpeed.ToString();
-                tPan.SetSpeed.Text = mSetSpeed.ToString();
+                tPan.CurSpeed.Text = mSelectedTrain.mCurSpeed.ToString();
+                tPan.CmdSpeed.Text = mSelectedTrain.mCmdSpeed.ToString();
+                tPan.SetSpeed.Text = mSelectedTrain.mSetSpeed.ToString();
 
-                if (mAutoMode)
+                if (mSelectedTrain.mAutoMode)
                 {
                     tPan.AutoMode.Content = "ON";
                     tPan.AutoMode.Background = new SolidColorBrush((Color)ColorConverter.ConvertFromString("#FF70D060"));
@@ -510,7 +303,7 @@ namespace TrainController
                     tPan.AutoMode.Background = new SolidColorBrush((Color)ColorConverter.ConvertFromString("#FFFF5050"));
                 }
 
-                if (mEmergencyBrakeStatus)
+                if (mSelectedTrain.mEmergencyBrakeStatus)
                 {
                     tPan.EmergencyBrake.Content = "ON";
                     tPan.EmergencyBrake.Background = new SolidColorBrush((Color)ColorConverter.ConvertFromString("#FF70D060"));
@@ -521,7 +314,7 @@ namespace TrainController
                     tPan.EmergencyBrake.Background = new SolidColorBrush((Color)ColorConverter.ConvertFromString("#FFFF5050"));
                 }
 
-                if (mServiceBrakeStatus)
+                if (mSelectedTrain.mServiceBrakeStatus)
                 {
                     tPan.ServiceBrake.Content = "ON";
                     tPan.ServiceBrake.Background = new SolidColorBrush((Color)ColorConverter.ConvertFromString("#FF70D060"));
@@ -532,149 +325,12 @@ namespace TrainController
                     tPan.ServiceBrake.Background = new SolidColorBrush((Color)ColorConverter.ConvertFromString("#FFFF5050"));
                 }
 
-                tPan.CmdAuthority.Text = mCmdAuthority.ToString();
-                tPan.CurAuthority.Text = mCurAuthority.ToString();
+                tPan.CmdAuthority.Text = mSelectedTrain.mCmdAuthority.ToString();
+                tPan.CurAuthority.Text = mSelectedTrain.mCurAuthority.ToString();
 
-                tPan.CurBeacon.Text = mBeacon;
+                tPan.CurBeacon.Text = mSelectedTrain.mBeacon;
 
                 tPan.Show();
-            }
-        }
-
-        public void setKp(int value)
-        {
-            mKp = value;
-
-            // Hardware Controls:
-            if(mControlType)
-            {
-                pi.WriteLine("j");
-                pi.WriteLine(mKp.ToString()+"\n");
-            }
-        }
-
-        public void setKi(int value)
-        {
-            mKi = value;
-
-            // Hardware Controls:
-            if (mControlType)
-            {
-                pi.WriteLine("k");
-                pi.WriteLine(mKi.ToString()+"\n");
-            }
-        }
-
-        public void setCmdSpeed(int value)
-        {
-            mCmdSpeed = value;
-
-            // Hardware Controls:
-            if (mControlType)
-            {
-                pi.WriteLine("v");
-                pi.WriteLine(mCmdSpeed.ToString()+"\n");
-            }
-        }
-
-        public void setSetSpeed(int value)
-        {
-            if (!mControlType)
-            {
-                if (value > mCmdSpeed)
-                {
-                    MessageBox.Show("Set Speed Shall Not Exceed Commanded Speed");
-                }
-                else
-                {
-                    mSetSpeed = value;
-                }
-            }
-            else
-            {
-                pi.WriteLine("b");
-                pi.WriteLine(value.ToString()+"\n");
-                string output = pi.ReadLine();
-
-                if (output == "tooHigh")
-                {
-                    MessageBox.Show("Set Speed Shall Not Exceed Commanded Speed");
-                }
-                else
-                {
-                    mSetSpeed = value;
-                }
-            }
-
-            SetSpeedBox.Text = mSetSpeed.ToString();
-        }
-
-        public void setCurSpeed(int value)
-        {
-            mCurSpeed = value;
-
-            // Hardware Controls:
-            if (mControlType)
-            {
-                pi.WriteLine("n");
-                pi.WriteLine(mCurSpeed.ToString()+"\n");
-            }
-        }
-
-        public void setCmdAuthority(int value)
-        {
-            mCmdAuthority = value;
-
-            // Hardware Controls:
-            if (mControlType)
-            {
-                pi.WriteLine("s");
-                pi.WriteLine(mCmdAuthority.ToString()+"\n");
-            }
-        }
-
-        public void setCurAuthority(int value)
-        {
-            mCurAuthority = value;
-
-            // Hardware Controls:
-            if (mControlType)
-            {
-                pi.WriteLine("d");
-                pi.WriteLine(mCurAuthority.ToString()+"\n");
-            }
-        }
-
-        public void setBeacon(string value)
-        {
-            // Software Controls:
-            if (!mControlType)
-            {
-                mBeacon = value;
-                Beacon.Text = "Nearest Beacon:\n" + value;
-            }
-
-            // Hardware Controls:
-            else
-            {
-                pi.WriteLine("f");
-                pi.WriteLine(value+"\n");
-                string output = pi.ReadLine();
-
-                mBeacon = output;
-                Beacon.Text = "Nearest Beacon:\n" + output;
-            }
-        }
-
-        public void setPower(int value)
-        {
-            mCurPower = value;
-
-            if (mControlType)
-            {
-                pi.WriteLine("p");
-                pi.WriteLine(value + "\n");
-                string output = pi.ReadLine();
             }
         }
 
@@ -684,258 +340,38 @@ namespace TrainController
             {
                 if (sender == SetSpeedBox)
                 {
-                    setSetSpeed(int.Parse(SetSpeedBox.Text));
+                    mSelectedTrain.setSetSpeed(int.Parse(SetSpeedBox.Text));
+                    SetSpeedBox.Text = mSelectedTrain.mSetSpeed.ToString();
                 }
             }
         }
 
         public void InitTimer()
         {
-            DispatcherTimer dispatcherTimer = new System.Windows.Threading.DispatcherTimer();
-            
-            if (!mControlType) dispatcherTimer.Tick += new EventHandler(SpeedUpdateSW);
-            else dispatcherTimer.Tick += new EventHandler(SpeedUpdateHW);
+            DispatcherTimer emergencyBrakeStatus = new DispatcherTimer();
 
-            dispatcherTimer.Interval = new TimeSpan(0,0,0,0,T);
-            dispatcherTimer.Start();
+            emergencyBrakeStatus.Tick += new EventHandler(checkUpdatedValues);
+
+            emergencyBrakeStatus.Interval = new TimeSpan(0, 0, 0, 0, T);
+            emergencyBrakeStatus.Start();
         }
 
-        public void SpeedUpdateSW(object sender, EventArgs e)
+        public void checkUpdatedValues(object sender, EventArgs e)
         {
-            if (mEmergencyBrakeStatus)
+            // Check for Emergency Brakes turned back off:
+            if(!mSelectedTrain.mEmergencyBrakeStatus)
             {
-                if (mCurSpeed == 0)
-                {
-                    mEmergencyBrakeStatus = false;
-                    EmergencyBrake.Content = "Emergency Brake\n         (OFF)";
-                    EmergencyBrake.Background = new SolidColorBrush((Color)ColorConverter.ConvertFromString("#FFFF5A5A"));
-                }
-                else
-                {
-                    mCmdSpeed = 0;
-                    CmdSpeed.Text = "Cmd Speed:\n" + mCmdSpeed + " mph";
-                    mSetSpeed = 0;
-                    SetSpeedBox.Text = mSetSpeed.ToString();
-
-                    mCurSpeed -= 1; // TODO: Replace with emergency brake deceleration!
-                    CurSpeed.Text = "Current Speed:\n" + mCurSpeed + " mph";
-                }
+                EmergencyBrake.Content = "Emergency Brake\n         (OFF)";
+                EmergencyBrake.Background = new SolidColorBrush((Color)ColorConverter.ConvertFromString("#FFFF5A5A"));
             }
-            else if (mServiceBrakeStatus)
-            {
-                if (mCurSpeed > 0)
-                {
-                    mCurSpeed--;  // TODO: Replace with service brake deceleration!
-                    CurSpeed.Text = "Current Speed:\n" + mCurSpeed + " mph";
-                }
-            }
-            else if (mAutoMode)
-            {
-                if (mCurSpeed < mCmdSpeed)
-                {
-                    mCurSpeed++;    // TODO: Replace with acceleration!
 
-                    if (mCurPower < Pmax)
-                    {
-                        Uk = Uk + (T / 1000) / 2 * (mCmdSpeed + mCurSpeed);
-                    }
-                    else
-                    {
-                        Uk = Uk;
-                    }
+            // Update Speed display:
+            CmdSpeed.Text = "Cmd Speed:\n" + mSelectedTrain.mCmdSpeed + " mph";
+            SetSpeedBox.Text = mSelectedTrain.mSetSpeed.ToString();
+            CurSpeed.Text = "Current Speed:\n" + mSelectedTrain.mCurSpeed + " mph";
 
-                    mCurPower = mKp * mCmdSpeed + mKi * Uk;
-                    CurPower.Text = "Power: " + mCurPower/1000 + " kW";
-
-                    CurSpeed.Text = "Current Speed:\n" + mCurSpeed + " mph";
-                }
-                else if (mCurSpeed > mCmdSpeed)
-                {
-                    mCurSpeed--;    // TODO: Replace with deceleration!
-
-                    if (mCurPower < Pmax)
-                    {
-                        Uk = Uk + (T / 1000) / 2 * (mCmdSpeed + mCurSpeed);
-                    }
-                    else
-                    {
-                        Uk = Uk;
-                    }
-
-                    mCurPower = -1 * (mKp * mCmdSpeed + mKi * Uk);
-                    CurPower.Text = "Power: " + mCurPower / 1000 + " kW";
-
-                    CurSpeed.Text = "Current Speed:\n" + mCurSpeed + " mph";
-                }
-                else
-                {
-                    mCurPower = 0;
-                    CurPower.Text = "Power: " + mCurPower / 1000 + " kW";
-                }
-            }
-            else
-            {
-                if (mCurSpeed < mSetSpeed)
-                {
-                    mCurSpeed++;    // TODO: Replace with acceleration!
-
-                    if (mCurPower < Pmax)
-                    {
-                        Uk = Uk + (T / 1000) / 2 * (mSetSpeed + mCurSpeed);
-                    }
-                    else
-                    {
-                        Uk = Uk;
-                    }
-
-                    mCurPower = mKp * mSetSpeed + mKi * Uk;
-                    CurPower.Text = "Power: " + mCurPower / 1000 + " kW";
-
-                    CurSpeed.Text = "Current Speed:\n" + mCurSpeed + " mph";
-                }
-                else if (mCurSpeed > mSetSpeed)
-                {
-                    mCurSpeed--;    // TODO: Replace with deceleration!
-
-                    if (mCurPower < Pmax)
-                    {
-                        Uk = Uk + (T / 1000) / 2 * (mSetSpeed + mCurSpeed);
-                    }
-                    else
-                    {
-                        Uk = Uk;
-                    }
-
-                    mCurPower = -1 * (mKp * mSetSpeed + mKi * Uk);
-                    CurPower.Text = "Power: " + mCurPower / 1000 + " kW";
-
-                    CurSpeed.Text = "Current Speed:\n" + mCurSpeed + " mph";
-                }
-                else
-                {
-                    mCurPower = 0;
-                    CurPower.Text = "Power: " + mCurPower / 1000 + " kW";
-                }
-            }
-        }
-
-        public void SpeedUpdateHW(object sender, EventArgs e)
-        {
-            if (mEmergencyBrakeStatus)
-            {
-                if (mCurSpeed == 0)
-                {
-                    mEmergencyBrakeStatus = false;
-                    EmergencyBrake.Content = "Emergency Brake\n         (OFF)";
-                    EmergencyBrake.Background = new SolidColorBrush((Color)ColorConverter.ConvertFromString("#FFFF5A5A"));
-                }
-                else
-                {
-                    mCmdSpeed = 0;
-                    CmdSpeed.Text = "Cmd Speed:\n" + mCmdSpeed + " mph";
-                    mSetSpeed = 0;
-                    SetSpeedBox.Text = mSetSpeed.ToString();
-
-                    mCurSpeed -= 1; // TODO: Replace with emergency brake deceleration!
-                    CurSpeed.Text = "Current Speed:\n" + mCurSpeed + " mph";
-                }
-            }
-            else if (mServiceBrakeStatus)
-            {
-                if (mCurSpeed > 0)
-                {
-                    mCurSpeed--;  // TODO: Replace with service brake deceleration!
-                    CurSpeed.Text = "Current Speed:\n" + mCurSpeed + " mph";
-                }
-            }
-            else if (mAutoMode)
-            {
-                if (mCurSpeed < mCmdSpeed)
-                {
-                    mCurSpeed++;    // TODO: Replace with acceleration!
-
-                    if (mCurPower < Pmax)
-                    {
-                        Uk = Uk + (T / 1000) / 2 * (mCmdSpeed + mCurSpeed);
-                    }
-                    else
-                    {
-                        Uk = Uk;
-                    }
-
-                    mCurPower = mKp * mCmdSpeed + mKi * Uk;
-                    CurPower.Text = "Power: " + mCurPower / 1000 + " kW";
-
-                    CurSpeed.Text = "Current Speed:\n" + mCurSpeed + " mph";
-                }
-                else if (mCurSpeed > mCmdSpeed)
-                {
-                    mCurSpeed--;    // TODO: Replace with deceleration!
-
-                    if (mCurPower < Pmax)
-                    {
-                        Uk = Uk + (T / 1000) / 2 * (mCmdSpeed + mCurSpeed);
-                    }
-                    else
-                    {
-                        Uk = Uk;
-                    }
-
-                    mCurPower = -1 * (mKp * mCmdSpeed + mKi * Uk);
-                    CurPower.Text = "Power: " + mCurPower / 1000 + " kW";
-
-                    CurSpeed.Text = "Current Speed:\n" + mCurSpeed + " mph";
-                }
-                else
-                {
-                    mCurPower = 0;
-                    CurPower.Text = "Power: " + mCurPower / 1000 + " kW";
-                }
-            }
-            else
-            {
-                if (mCurSpeed < mSetSpeed)
-                {
-                    mCurSpeed++;    // TODO: Replace with acceleration!
-
-                    if (mCurPower < Pmax)
-                    {
-                        Uk = Uk + (T / 1000) / 2 * (mSetSpeed + mCurSpeed);
-                    }
-                    else
-                    {
-                        Uk = Uk;
-                    }
-
-                    mCurPower = mKp * mSetSpeed + mKi * Uk;
-                    CurPower.Text = "Power: " + mCurPower / 1000 + " kW";
-
-                    CurSpeed.Text = "Current Speed:\n" + mCurSpeed + " mph";
-                }
-                else if (mCurSpeed > mSetSpeed)
-                {
-                    mCurSpeed--;    // TODO: Replace with deceleration!
-
-                    if (mCurPower < Pmax)
-                    {
-                        Uk = Uk + (T / 1000) / 2 * (mSetSpeed + mCurSpeed);
-                    }
-                    else
-                    {
-                        Uk = Uk;
-                    }
-
-                    mCurPower = -1 * (mKp * mSetSpeed + mKi * Uk);
-                    CurPower.Text = "Power: " + mCurPower / 1000 + " kW";
-
-                    CurSpeed.Text = "Current Speed:\n" + mCurSpeed + " mph";
-                }
-                else
-                {
-                    mCurPower = 0;
-                    CurPower.Text = "Power: " + mCurPower / 1000 + " kW";
-                }
-            }
+            // Update Power display:
+            CurPower.Text = "Power: " + mSelectedTrain.mCurPower / 1000 + " kW";
         }
     }
 }
